Safari-609.1.11.1/LayoutTests/fast/files/file-reader-event-listener-expected.txt

Test that FileReader's event listener works.
Received loadstart event
Received foo event
PASS event.bubbles is false
PASS event.cancelable is false
PASS event.defaultPrevented is false
Received progress event
Received load event
Received loadend event
PASS successfullyParsed is true

TEST COMPLETE
